feat(refs: DPLAN-14556, , #AB22376)  Extract table functionality into reusable OrganisationTable component
  - Create new OrganisationTable component from DpAddOrganisationList table logic
  - Add resourceType prop to support different data types (InvitableToeb, Member)
  - Move all table-related functionality to OrganisationTable:
    * API calls and data loading (getInstitutionsWithContacts)
    * Search and filter UI with FilterFlyout integration
    * Pagination with DpPager component
    * Vuex store integration for all required modules
  - Simplify DpAddOrganisationList to wrapper component with only action buttons
  - Update administrationMemberList bundle to register required Vuex stores
  - Enable component reusability across different contexts (Vue and Twig)
