+++ client/js/bundles/procedure/administrationMemberList.js
@@ -12,13 +12,27 @@
  */
 import { addFormHiddenField, removeFormHiddenField } from '../../lib/core/libs/FormActions'
 import { DpContextualHelp } from '@demos-europe/demosplan-ui'
+import FilterFlyoutStore from '@DpJs/store/procedure/FilterFlyout'
 import { initialize } from '@DpJs/InitVue'
+import OrganisationTable from '@DpJs/components/procedure/admin/InstitutionTagManagement/OrganisationTable.vue'
 
 const components = {
-  DpContextualHelp
+  DpContextualHelp,
+  OrganisationList: OrganisationTable
 }
 
-initialize(components).then(() => {
+const apiStores = [
+  'InvitableToeb',
+  'InstitutionTag',
+  'InstitutionTagCategory',
+  'InstitutionLocationContact'
+]
+
+const stores = {
+  FilterFlyout: FilterFlyoutStore
+}
+
+initialize(components, stores, apiStores).then(() => {
   const pdfExport = () => {
     const action = document.procedureForm.action
     const target = document.procedureForm.target
